+++ src/main/java/com/passingtest/service/QuestionService.java
@@ -5,6 +5,7 @@ import com.passingtest.repository.QuestionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,7 +14,7 @@ public class QuestionService {
     @Autowired
     QuestionRepository questionRepository;
 
-    public Question getQuestionById(int id) {
+    public Question getQuestionById(BigInteger id) {
         return questionRepository.findById(id).get();
     }
 
@@ -21,7 +22,7 @@ public class QuestionService {
         questionRepository.save(question);
     }
 
-    public List<Question> getQuestionsByTestId(Integer testId) {
+    public List<Question> getQuestionsByTestId(BigInteger testId) {
         List<Question> questions = new ArrayList<Question>();
         questionRepository.getQuestionsByTestId(testId).forEach(question1 -> questions.add(question1));
         return questions;
